requirements.txt
flask
python-docx
pillow
pytesseract